feat(routes): /storage
